feat(irq): implement aarch64 interrupt handling and devicetree routing
- Implement `_el1h_irq_handler` in assembly to save/restore context and call the dispatcher.
- Implement `global_irq_dispatch` and `dispatch_irq` with lock-dropping to prevent re-entrancy deadlocks.
- Implement `resolve_virq` using DTSpec-compliant interrupt-parent resolution logic.
- Enhance `DeviceTree` with a `BTreeMap` for O(log n) node lookups by phandle.
- Add `enable_irq_fiq` to unmask PSTATE DAIF bits.
- Update BCM2836 L1 interrupt controller driver to use new `InterruptSpecifier` types.

+++ src/kernel/devicetree/interrupts.rs
@@ -40,11 +40,11 @@ impl Interrupts {
         Self { raw_data: data }
     }
 
-    pub fn iter(&self, interrupt_cells: usize) -> impl Iterator<Item = InterruptSpecifier> {
+    pub fn iter(&self, interrupt_cells: u32) -> impl Iterator<Item = InterruptSpecifier> {
         assert!(interrupt_cells > 0);
 
         self.raw_data
-            .chunks_exact(interrupt_cells)
+            .chunks_exact(interrupt_cells as usize)
             .map(|chunk| InterruptSpecifier(chunk.to_vec()))
     }
 }
